services/orchestrator-agent/requirements.txt
discord.py>=2.0.0
docker==7.1.0
python-dotenv>=0.19.0
colorama==0.4.6
psutil>=5.8.0
matplotlib>=3.5.0
asyncio>=3.4.3
aiohttp>=3.8.0
aiomysql>=0.1.1
aiosqlite>=0.17.0
cryptography>=36.0.0
pillow>=9.0.0
typing-extensions>=4.0.0
requests>=2.26.0